P4ADEV-2134 removed interfaces InstallmentUpdateService and PrimaryOrgInstallmentService
